test(discriminators): rearrange test files and test cases
