+++ src/__tests__/composeWithMongooseDiscriminators-test.js
@@ -1,13 +1,6 @@
 /* @flow */
 
-import {
-  graphql,
-  InputTypeComposer,
-  SchemaComposer,
-  schemaComposer,
-  TypeComposer,
-  InterfaceTypeComposer,
-} from 'graphql-compose';
+import { InputTypeComposer, schemaComposer, TypeComposer } from 'graphql-compose';
 import { getCharacterModels } from '../__mocks__/characterModels';
 import { MovieModel } from '../__mocks__/movieModel';
 import { composeWithMongooseDiscriminators } from '../composeWithMongooseDiscriminators';
@@ -30,15 +23,10 @@ describe('composeWithMongooseDiscriminators ->', () => {
       );
     });
 
-    it('should return a TypeComposer as childTC', () => {
+    it('should return a TypeComposer as childTC on discriminator() call', () => {
       expect(
         composeWithMongooseDiscriminators(CharacterModel).discriminator(PersonModel)
       ).toBeInstanceOf(TypeComposer);
-    });
-
-    it('should have an interface, accessed with getDInterface', () => {
-      const cDTC = composeWithMongooseDiscriminators(CharacterModel);
-      expect(cDTC.getDInterface()).toBeInstanceOf(InterfaceTypeComposer);
     });
   });
 
@@ -71,193 +59,6 @@ describe('composeWithMongooseDiscriminators ->', () => {
 
       expect(itc.isRequired('name')).toBe(true);
       expect(itc.isRequired('friends')).toBe(true);
-    });
-  });
-
-  describe('DInterface', () => {
-    it('should have same field names as baseModel used to create it', () => {
-      const baseDTC = composeWithMongooseDiscriminators(CharacterModel);
-      expect(baseDTC.getFieldNames()).toEqual(
-        expect.arrayContaining(Object.keys(baseDTC.getDInterface().getFields()))
-      );
-    });
-
-    it('should have field names synced with the baseTC', () => {
-      const baseDTC = composeWithMongooseDiscriminators(CharacterModel);
-
-      expect(baseDTC.getFieldNames()).toEqual(Object.keys(baseDTC.getDInterface().getFields()));
-
-      beforeAll(() =>
-        baseDTC.addFields({
-          field1: 'String',
-          field2: 'String',
-        }));
-
-      expect(baseDTC.getFieldNames()).toEqual(Object.keys(baseDTC.getDInterface().getFields()));
-    });
-  });
-
-  describe('DiscriminatorTypeComposer', () => {
-    it('should have as interface DInterface', () => {
-      const baseDTC = composeWithMongooseDiscriminators(CharacterModel);
-      expect(baseDTC.hasInterface(baseDTC.getDInterface())).toBeTruthy();
-    });
-
-    describe('hasChildTC(DName)', () => {
-      const baseDTC = composeWithMongooseDiscriminators(CharacterModel);
-      const personModel = baseDTC.discriminator(PersonModel);
-
-      it('should check and return boolean if childDTC is available', () => {
-        expect(baseDTC.hasChildTC(personModel.getTypeName())).toBeTruthy();
-      });
-
-      it('should be falsified as childDTC not found', () => {
-        expect(baseDTC.hasChildTC('NOT_AVAILABLE')).toBeFalsy();
-      });
-    });
-
-    describe('addFields(newFields)', () => {
-      const characterDTC = composeWithMongooseDiscriminators(CharacterModel);
-      const personTC = characterDTC.discriminator(PersonModel);
-      const droidTC = characterDTC.discriminator(DroidModel);
-      const newFields = {
-        field1: 'String',
-        field2: 'String',
-      };
-
-      beforeAll(() => {
-        characterDTC.addFields(newFields);
-      });
-
-      it('should add fields to baseTC', () => {
-        expect(characterDTC.getFieldNames()).toEqual(
-          expect.arrayContaining(Object.keys(newFields))
-        );
-      });
-
-      it('should add fields to DInterface', () => {
-        expect(Object.keys(characterDTC.getDInterface().getFields())).toEqual(
-          expect.arrayContaining(Object.keys(newFields))
-        );
-      });
-
-      it('should add fields to childTC', () => {
-        expect(personTC.getFieldNames()).toEqual(expect.arrayContaining(Object.keys(newFields)));
-        expect(droidTC.getFieldNames()).toEqual(expect.arrayContaining(Object.keys(newFields)));
-      });
-    });
-
-    describe('removeField()', () => {
-      const characterDTC = composeWithMongooseDiscriminators(CharacterModel);
-      const personTC = characterDTC.discriminator(PersonModel);
-      const droidTC = characterDTC.discriminator(DroidModel);
-      const field = 'friends';
-
-      beforeAll(() => {
-        characterDTC.removeField(field);
-      });
-
-      it('should remove fields from baseTC', () => {
-        expect(characterDTC.hasField(field)).toBeFalsy();
-      });
-
-      it('should remove fields from DInterface', () => {
-        expect(characterDTC.getDInterface().getFields()[field]).toBeFalsy();
-      });
-
-      it('should remove fields from childTC', () => {
-        expect(personTC.hasField(field)).toBeFalsy();
-        expect(droidTC.hasField(field)).toBeFalsy();
-      });
-    });
-
-    describe('extendFields(fieldName, extensionField)', () => {
-      const characterDTC = composeWithMongooseDiscriminators(CharacterModel);
-      const personTC = characterDTC.discriminator(PersonModel);
-      const droidTC = characterDTC.discriminator(DroidModel);
-      const fieldName = 'kind';
-      const fieldExtension = {
-        type: 'String',
-        description: 'Hello I am changed',
-      };
-
-      beforeAll(() => {
-        characterDTC.extendField(fieldName, fieldExtension);
-      });
-
-      it('should extend field on baseTC', () => {
-        expect(characterDTC.getFieldType(fieldName).toString()).toEqual(graphql.GraphQLString.name);
-
-        expect((characterDTC.getField(fieldName): any).description).toEqual(
-          fieldExtension.description
-        );
-      });
-
-      it('should extend field type on DInterface', () => {
-        expect(characterDTC.getDInterface().getFields()[fieldName]).toBeTruthy();
-        expect(
-          characterDTC
-            .getDInterface()
-            .getFieldType(fieldName)
-            .toString()
-        ).toEqual(fieldExtension.type);
-      });
-
-      it('should extend field on childTC', () => {
-        expect(personTC.getFieldType(fieldName).toString()).toEqual(graphql.GraphQLString.name);
-
-        expect((personTC.getField(fieldName): any).description).toEqual(fieldExtension.description);
-
-        expect(droidTC.getFieldType(fieldName).toString()).toEqual(graphql.GraphQLString.name);
-
-        expect((droidTC.getField(fieldName): any).description).toEqual(fieldExtension.description);
-      });
-    });
-
-    describe('discriminator()', () => {
-      let sc;
-      let characterDTC;
-
-      beforeEach(() => {
-        sc = new SchemaComposer();
-        characterDTC = composeWithMongooseDiscriminators(CharacterModel, {
-          customizationOptions: { schemaComposer: sc },
-        });
-      });
-
-      it('should return an instance of TypeComposer as childTC', () => {
-        /*
-          Test keeps on failing, FIXME: Recheck
-          Expected constructor: TypeComposer
-          Received constructor: TypeComposer
-          Received value: {"gqType": "Person"} */
-        // expect(characterDTC.discriminator(PersonModel)).toBeInstanceOf(TypeComposer);
-        // expect(characterDTC.discriminator(DroidModel)).toBeInstanceOf(TypeComposer);
-      });
-
-      it('should register itself in childTC(childTCs) array', () => {
-        const childTC = characterDTC.discriminator(DroidModel);
-        expect(characterDTC.hasChildTC(childTC.getTypeName())).toBeTruthy();
-      });
-
-      it('should apply filters passed', () => {
-        const tc = characterDTC.discriminator(PersonModel, {
-          fields: {
-            remove: ['dob', 'starShips'],
-          },
-        });
-
-        expect(tc.getFieldNames()).not.toEqual(expect.arrayContaining(['dob', 'starShips']));
-      });
-    });
-  });
-
-  describe('DiscriminatorTypes', () => {
-    it('should have as an interface DInterface', () => {
-      const baseDTC = composeWithMongooseDiscriminators(CharacterModel);
-      expect(baseDTC.discriminator(DroidModel).getInterfaces()).toEqual(
-        expect.arrayContaining(Array.of(baseDTC.getDInterface()))
-      );
     });
   });
 });
